// npm/sdk/src/types/lumen/gateway/v1/tx.ts
export interface MsgRegisterGateway {
  operator: string;
  payout: string;
  metadata: string;
}

export interface MsgRegisterGatewayResponse {
  id: number;
}

export interface MsgUpdateGateway {
  operator: string;
  gatewayId: number;
  payout: string | undefined;
  metadata: string | undefined;
  active: boolean | undefined;
}

export interface MsgUpdateGatewayResponse {
}

export interface MsgUpdateParams {
  authority: string;
  params: Params | undefined;
}

export interface MsgUpdateParamsResponse {
}

export interface MsgCreateContract {
  client: string;
  gatewayId: number;
  priceUlmn: number;
  storageGbPerMonth: number;
  networkGbPerMonth: number;
  monthsTotal: number;
  metadata: string;
}

export interface MsgCreateContractResponse {
  contractId: number;
}

export interface MsgClaimPayment {
  operator: string;
  contractId: number;
}

export interface MsgClaimPaymentResponse {
  paidUlmn: string;
}

export interface MsgCancelContract {
  client: string;
  contractId: number;
}

export interface MsgCancelContractResponse {
  refundedUlmn: string;
}

export interface MsgFinalizeContract {
  finalizer: string;
  contractId: number;
}
// ...
export const MsgCreateContract: MessageFns<MsgCreateContract> = {
  encode(message: MsgCreateContract, writer: BinaryWriter = new BinaryWriter()): BinaryWriter {
    if (message.client !== "") {
      writer.uint32(10).string(message.client);
    }
    if (message.gatewayId !== 0) {
      writer.uint32(16).uint64(message.gatewayId);
    }
    if (message.priceUlmn !== 0) {
      writer.uint32(24).uint64(message.priceUlmn);
    }
    if (message.storageGbPerMonth !== 0) {
      writer.uint32(32).uint64(message.storageGbPerMonth);
    }
    if (message.networkGbPerMonth !== 0) {
      writer.uint32(40).uint64(message.networkGbPerMonth);
    }
    if (message.monthsTotal !== 0) {
      writer.uint32(48).uint32(message.monthsTotal);
    }
    if (message.metadata !== "") {
      writer.uint32(58).string(message.metadata);
    }
    return writer;
  },

  decode(input: BinaryReader | Uint8Array, length?: number): MsgCreateContract {
    const reader = input instanceof BinaryReader ? input : new BinaryReader(input);
    let end = length === undefined ? reader.len : reader.pos + length;
    const message = createBaseMsgCreateContract();
    while (reader.pos < end) {
      const tag = reader.uint32();
      switch (tag >>> 3) {
        case 1: {
          if (tag !== 10) {
            break;
          }

          message.client = reader.string();
          continue;
        }
        case 2: {
          if (tag !== 16) {
            break;
          }

          message.gatewayId = longToNumber(reader.uint64());
          continue;
        }
        case 3: {
          if (tag !== 24) {
            break;
          }

          message.priceUlmn = longToNumber(reader.uint64());
          continue;
        }
        case 4: {
          if (tag !== 32) {
            break;
          }

          message.storageGbPerMonth = longToNumber(reader.uint64());
          continue;
        }
        case 5: {
          if (tag !== 40) {
            break;
          }

          message.networkGbPerMonth = longToNumber(reader.uint64());
          continue;
        }
        case 6: {
          if (tag !== 48) {
            break;
          }

          message.monthsTotal = reader.uint32();
          continue;
        }
        case 7: {
          if (tag !== 58) {
            break;
          }

          message.metadata = reader.string();
          continue;
        }
      }
      if ((tag & 7) === 4 || tag === 0) {
        break;
      }
      reader.skip(tag & 7);
    }
    return message;
  },

  fromJSON(object: any): MsgCreateContract {
    return {
      client: isSet(object.client) ? globalThis.String(object.client) : "",
      gatewayId: isSet(object.gatewayId) ? globalThis.Number(object.gatewayId) : 0,
      priceUlmn: isSet(object.priceUlmn) ? globalThis.Number(object.priceUlmn) : 0,
      storageGbPerMonth: isSet(object.storageGbPerMonth) ? globalThis.Number(object.storageGbPerMonth) : 0,
      networkGbPerMonth: isSet(object.networkGbPerMonth) ? globalThis.Number(object.networkGbPerMonth) : 0,
      monthsTotal: isSet(object.monthsTotal) ? globalThis.Number(object.monthsTotal) : 0,
      metadata: isSet(object.metadata) ? globalThis.String(object.metadata) : "",
    };
  },

  toJSON(message: MsgCreateContract): unknown {
    const obj: any = {};
    if (message.client !== "") {
      obj.client = message.client;
    }
    if (message.gatewayId !== 0) {
      obj.gatewayId = Math.round(message.gatewayId);
    }
    if (message.priceUlmn !== 0) {
      obj.priceUlmn = Math.round(message.priceUlmn);
    }
    if (message.storageGbPerMonth !== 0) {
      obj.storageGbPerMonth = Math.round(message.storageGbPerMonth);
    }
    if (message.networkGbPerMonth !== 0) {
      obj.networkGbPerMonth = Math.round(message.networkGbPerMonth);
    }
    if (message.monthsTotal !== 0) {
      obj.monthsTotal = Math.round(message.monthsTotal);
    }
    if (message.metadata !== "") {
      obj.metadata = message.metadata;
    }
    return obj;
  },

  create<I extends Exact<DeepPartial<MsgCreateContract>, I>>(base?: I): MsgCreateContract {
    return MsgCreateContract.fromPartial(base ?? ({} as any));
  },
  fromPartial<I extends Exact<DeepPartial<MsgCreateContract>, I>>(object: I): MsgCreateContract {
    const message = createBaseMsgCreateContract();
    message.client = object.client ?? "";
    message.gatewayId = object.gatewayId ?? 0;
    message.priceUlmn = object.priceUlmn ?? 0;
    message.storageGbPerMonth = object.storageGbPerMonth ?? 0;
    message.networkGbPerMonth = object.networkGbPerMonth ?? 0;
    message.monthsTotal = object.monthsTotal ?? 0;
    message.metadata = object.metadata ?? "";
    return message;
  },
};
// ...
export const MsgClaimPayment: MessageFns<MsgClaimPayment> = {
  encode(message: MsgClaimPayment, writer: BinaryWriter = new BinaryWriter()): BinaryWriter {
    if (message.operator !== "") {
      writer.uint32(10).string(message.operator);
    }
    if (message.contractId !== 0) {
      writer.uint32(16).uint64(message.contractId);
    }
    return writer;
  },

  decode(input: BinaryReader | Uint8Array, length?: number): MsgClaimPayment {
    const reader = input instanceof BinaryReader ? input : new BinaryReader(input);
    let end = length === undefined ? reader.len : reader.pos + length;
    const message = createBaseMsgClaimPayment();
    while (reader.pos < end) {
      const tag = reader.uint32();
      switch (tag >>> 3) {
        case 1: {
          if (tag !== 10) {
            break;
          }

          message.operator = reader.string();
          continue;
        }
        case 2: {
          if (tag !== 16) {
            break;
          }

          message.contractId = longToNumber(reader.uint64());
          continue;
        }
      }
      if ((tag & 7) === 4 || tag === 0) {
        break;
      }
      reader.skip(tag & 7);
    }
    return message;
  },

  fromJSON(object: any): MsgClaimPayment {
    return {
      operator: isSet(object.operator) ? globalThis.String(object.operator) : "",
      contractId: isSet(object.contractId) ? globalThis.Number(object.contractId) : 0,
    };
  },

  toJSON(message: MsgClaimPayment): unknown {
    const obj: any = {};
    if (message.operator !== "") {
      obj.operator = message.operator;
    }
    if (message.contractId !== 0) {
      obj.contractId = Math.round(message.contractId);
    }
    return obj;
  },

  create<I extends Exact<DeepPartial<MsgClaimPayment>, I>>(base?: I): MsgClaimPayment {
    return MsgClaimPayment.fromPartial(base ?? ({} as any));
  },
  fromPartial<I extends Exact<DeepPartial<MsgClaimPayment>, I>>(object: I): MsgClaimPayment {
    const message = createBaseMsgClaimPayment();
    message.operator = object.operator ?? "";
    message.contractId = object.contractId ?? 0;
    return message;
  },
};
// ...
export const MsgCancelContract: MessageFns<MsgCancelContract> = {
  encode(message: MsgCancelContract, writer: BinaryWriter = new BinaryWriter()): BinaryWriter {
    if (message.client !== "") {
      writer.uint32(10).string(message.client);
    }
    if (message.contractId !== 0) {
      writer.uint32(16).uint64(message.contractId);
    }
    return writer;
  },

  decode(input: BinaryReader | Uint8Array, length?: number): MsgCancelContract {
    const reader = input instanceof BinaryReader ? input : new BinaryReader(input);
    let end = length === undefined ? reader.len : reader.pos + length;
    const message = createBaseMsgCancelContract();
    while (reader.pos < end) {
      const tag = reader.uint32();
      switch (tag >>> 3) {
        case 1: {
          if (tag !== 10) {
            break;
          }

          message.client = reader.string();
          continue;
        }
        case 2: {
          if (tag !== 16) {
            break;
          }

          message.contractId = longToNumber(reader.uint64());
          continue;
        }
      }
      if ((tag & 7) === 4 || tag === 0) {
        break;
      }
      reader.skip(tag & 7);
    }
    return message;
  },

  fromJSON(object: any): MsgCancelContract {
    return {
      client: isSet(object.client) ? globalThis.String(object.client) : "",
      contractId: isSet(object.contractId) ? globalThis.Number(object.contractId) : 0,
    };
  },

  toJSON(message: MsgCancelContract): unknown {
    const obj: any = {};
    if (message.client !== "") {
      obj.client = message.client;
    }
    if (message.contractId !== 0) {
      obj.contractId = Math.round(message.contractId);
    }
    return obj;
  },

  create<I extends Exact<DeepPartial<MsgCancelContract>, I>>(base?: I): MsgCancelContract {
    return MsgCancelContract.fromPartial(base ?? ({} as any));
  },
  fromPartial<I extends Exact<DeepPartial<MsgCancelContract>, I>>(object: I): MsgCancelContract {
    const message = createBaseMsgCancelContract();
    message.client = object.client ?? "";
    message.contractId = object.contractId ?? 0;
    return message;
  },
};
// ...
export const MsgFinalizeContract: MessageFns<MsgFinalizeContract> = {
  encode(message: MsgFinalizeContract, writer: BinaryWriter = new BinaryWriter()): BinaryWriter {
    if (message.finalizer !== "") {
      writer.uint32(10).string(message.finalizer);
    }
    if (message.contractId !== 0) {
      writer.uint32(16).uint64(message.contractId);
    }
    return writer;
  },

  decode(input: BinaryReader | Uint8Array, length?: number): MsgFinalizeContract {
    const reader = input instanceof BinaryReader ? input : new BinaryReader(input);
    let end = length === undefined ? reader.len : reader.pos + length;
    const message = createBaseMsgFinalizeContract();
    while (reader.pos < end) {
      const tag = reader.uint32();
      switch (tag >>> 3) {
        case 1: {
          if (tag !== 10) {
            break;
          }

          message.finalizer = reader.string();
          continue;
        }
        case 2: {
          if (tag !== 16) {
            break;
          }

          message.contractId = longToNumber(reader.uint64());
          continue;
        }
      }
      if ((tag & 7) === 4 || tag === 0) {
        break;
      }
      reader.skip(tag & 7);
    }
    return message;
  },

  fromJSON(object: any): MsgFinalizeContract {
    return {
      finalizer: isSet(object.finalizer) ? globalThis.String(object.finalizer) : "",
      contractId: isSet(object.contractId) ? globalThis.Number(object.contractId) : 0,
    };
  },

  toJSON(message: MsgFinalizeContract): unknown {
    const obj: any = {};
    if (message.finalizer !== "") {
      obj.finalizer = message.finalizer;
    }
    if (message.contractId !== 0) {
      obj.contractId = Math.round(message.contractId);
    }
    return obj;
  },

  create<I extends Exact<DeepPartial<MsgFinalizeContract>, I>>(base?: I): MsgFinalizeContract {
    return MsgFinalizeContract.fromPartial(base ?? ({} as any));
  },
  fromPartial<I extends Exact<DeepPartial<MsgFinalizeContract>, I>>(object: I): MsgFinalizeContract {
    const message = createBaseMsgFinalizeContract();
    message.finalizer = object.finalizer ?? "";
    message.contractId = object.contractId ?? 0;
    return message;
  },
};